--- app/login/page.tsx
@@ -1,5 +1,0 @@
-import LoginForm from '@/app/ui/login-form';
-
-export default function Page() {
-  return <LoginForm />;
-}
